CHROMIUM: gobi: Don't leak urb data buffers
We make a copy of the data in each skb we send, and pass it off to the usb
subsystem.  Unfortunately, we weren't setting the "please free the buffer
when you don't need it anymore" flag, and we weren't freeing it ourselves,
so we leaked all of the urb data buffers.

Now we set the flag, and we don't leak the buffers.

BUG=None
TEST=network_3GSmokeTest

Signed-off-by: ttuttle <ttuttle@chromium.org>

Change-Id: Ie624dcc074485e374ebdd82d15396a5741cecb4e
Reviewed-on: http://gerrit.chromium.org/gerrit/4838
Reviewed-by: Jason Glasgow <jglasgow@chromium.org>
Tested-by: ttuttle <ttuttle@chromium.org>