msm: camera: Change max number of application buffers allowed.
During AXI configuration, the buffer addresses are
stored in a msm_pmem_region array of size 16. This
limits the maximum number of buffers that can be
used by the application to 16. But in ZSL mode,
the application can allocate more than 16 buffers.
Increase the maximum number of buffers and
dynamically allocate memory for this structure.

Change-Id: Ida227b0949dd7344c8230c58ef7d345463c69ca7
Signed-off-by: Kiran Kumar H N <hurlisal@codeaurora.org>