mmc: msm_sdcc: put regulator in HPM before enabling it.
If regulator is enabled when it's in LPM (Low Power Mode),
it takes ~400us to enable it where as if regulator is
enabled when it's in HPM mode, it take < 200us to enable
it. This means immediately after enabling the regulator,
it's usable as it was already in HPM mode.

Change-Id: Iadf3960ef030b2edb4d93ff34a68c62cbdb0445a
Signed-off-by: Subhash Jadavani <subhashj@codeaurora.org>