ASoC: dsp: Scan only active playback or capture stream
Digital Audio Interface(DAI) can be defined to support playback
and capture. When there is a PCM capture running on bi-drectional
CPU DAI, soc-dsp framework can errorneously start backend which
is connected to the same CPU DAI on the playback path if platform
mixer command is issued to join playback frontend to given backend.

CRs-fixed: 298954
Change-Id: I5d0dafd049e1108506f114870bd2a42dadba8fe8
Signed-off-by: Patrick Lai <plai@codeaurora.org>