board: msm8x60: allow hw to stabilize on resume
Waiting for I2C-slave (touchscreen) stabilization. Some
slaves need this delay after their RESET line is taken out of reset, others
need it when power is applied if there is no RESET line.  In touchscreen
case, this appears to be added after regulator is turned on as well since
there is no explicit RESET line being used.  The time for I2C
slave-stabilization varies by slave and most of the times, it's not
explicitly specified in the specs. So it has to change from board to board.

CRs-fixed: 297832
Change-Id: Ib3dd37ff82e23c30da67f2d5043ff436a01bb0fd
Signed-off-by: Amy Maloche <amaloche@codeaurora.org>