regulator: pm8921-regulator: add hack to boost LDO 4 voltage to 2.0 V
PMIC 8921 regulator LDO 4 should normally be set to 1.8 V.  However,
for PMIC revisions 1.1 and below, this voltage is insufficient for
internal PMIC circuitry.  This results in some SSBI registers in the
PMIC not reading back properly.  Therefore, add a hack which raises
the voltage set for LDO 4 to 2.0 V to fix the issue.

Change-Id: I38e88250307d262f544129fcc31c4021a6ba2d3f
Signed-off-by: David Collins <collinsd@codeaurora.org>