qup_i2c: Fix state check in polling code.
If QUP core is in PAUSE state then the state check returns true
for any other state incorrectly. Also, if the core is in RUN state
and it's being put in RESET, then RESET state check may return true
even if it's not out of RUN (and if VALID bit is set).
State mask is used to verify what state QUP core is in to avoid this
problem.

Change-Id: I09a6bc08141d9bca502aa81f7ed406f170eb61d5
Signed-off-by: Sagar Dharia <sdharia@codeaurora.org>