qup_i2c: Fix state check in polling code.
If QUP core is in PAUSE state then the state check returns true
for any other state incorrectly. Also, if the core is in RUN state
and it's being put in RESET, then RESET state check may return true
even if it's not out of RUN (and if VALID bit is set).
State mask is used to verify what state QUP core is in to avoid this
problem.

Signed-off-by: Sagar Dharia <sdharia@codeaurora.org>
(cherry picked from commit 1abd73dcd6048907fd05370807d79d18746a7766)

Change-Id: If70acd72a3da8af3c7ee04e7e24e96a26be05c40
Signed-off-by: Oguri, Srivalli <oguri@codeaurora.org>