mmc: msm_sdcc: Configure sdcc2 and sdcc5 GPIO's to always PULL_UP.
As per the SD spec the CMD and DAT lines need to be set
to PULL_UP, but we set it to PULL_NONE during sleep. This was
causing the client to get into an error state during sleep/wakeup
sequence on rare scenarios. Now setting them to PULL_UP on
both active and sleep state.

Change-Id: I13bfba4acbd63685a6e1bb52a0fbae8e7dcca872
Signed-off-by: Venkat Gopalakrishnan <venkatg@codeaurora.org>