CHROMIUM: audio: cirrus: add support for CS421x
This update includes the changes necessary for supporting the CS421x
family of codecs. Previously this file only supported the CS420x
family of codecs.

This file also contains initialization verbs to correct several issues
in both the CS420x and CS421x hardware.

NOTE: a version of this patch has been uploaded upstream to ALSA.
However, in order to use the upstreamed patch we need to be at ALSO version
1.0.25. We are currently at version 1.0.24 and therefore taking this out
of tree patch for now.

TEST=plug in headphones. Go to an audio site such as zombo.com and make
sure audio is heard.
BUG=chrome-os-partner:5290

Change-Id: I370a19ce0b84534f33e18e5fe1f5f6fa799394ae
Signed-off-by: Tim Howe <tim.howe@cirrus.com>
Reviewed-on: http://gerrit.chromium.org/gerrit/5663
Tested-by: Sameer Nanda <snanda@chromium.org>
Reviewed-by: Todd Broch <tbroch@chromium.org>