power: pm8921-charger: Fix VBUS reverse voltage
A voltage of 4.2 volts was observed on the vbus pin even after the
usb was removed.

The reason was the buck in the charger module was behaving in reverse
causing voltage to appear on vbus line.

Fix this by writing a value of 0xF1 to CHG_BUCK_CNTRL_3 register.

CRs-fixed: 299784
Change-Id: Ib2d248fcfd98925d1a9e14f92e4a6e61f28ab4a8
Signed-off-by: Abhijeet Dharmapurikar <adharmap@codeaurora.org>