CHROMIUM: config: enable large block devices on ARM
This change allows to mount USB key formatted with ext4 filesystem on
ARM based configuration (this is already working on x86).

CONFIG_LBDAF normally enables the support for 2TB+ block devices,
but when an ext4 filesystem has the "huge_file" feature enabled (which is
the default in mkfs.ext4), it requires this option to be mounted
read-write.

The GFS2_FS configuration symbol is now common to all our configuration file,
so it moves automatically to config.common.chromeos.

Signed-off-by: Vincent Palatin <vpalatin@chromium.org>

BUG=chromium-os:19106
TEST=mount an EXT4 formatted USB key from the command line

Change-Id: If97eac4f880cc413e42c40ac71a6e48c92019737
Reviewed-on: http://gerrit.chromium.org/gerrit/5900
Reviewed-by: Will Drewry <wad@chromium.org>
Tested-by: Vincent Palatin <vpalatin@chromium.org>