ARM: Perfevents: Remove erroneous constraint on number of active events
The original intent was to restrict the number of active events to number
of available counters at init. But this code had the side effect of events
needelessly reserving a slot even when they couldn't be armed on the PMU.

Eventually this resource reservation code will be replaced entirely.
For now, we lazily deny event allocation just before arming them on
the PMU.

Change-Id: I985a078660371f208e92052ac7dc5fbced72e22c
Signed-off-by: Ashwin Chaugule <ashwinc@codeaurora.org>