ASoC: WCD9310: Turn on Line3 and Line4 DAC's for speakers
In Tabla 1.0, negative signal from RX3(connected lineout1 DAC)
and RX5(connected to lineout2 DAC) cannot be sent to lineout3
and lineout4 DAC's respectively. For differential speakers to
work correctly, lineout3 and lineout4 DAC still need to be
turned on even if no audio signal is sent to them.

Change-Id: I893f1ef579fddbea33f3e379c8fbfc0f8e951bb4
CRs-Fixed: 298454, 300113
Signed-off-by: Kiran Kandi <kkandi@codeaurora.org>