CHROMIUM: Input: synaptics - process finger (<=5) transitions
Synaptics image sensor touchpads track up to 5 fingers, but only report 2.
They use a special "TYPE=2" (AGM-CONTACT) packet type that reports
the number of tracked fingers and which finger is reported in the SGM
and AGM packets.

With this new packet type, it is possible to tell userspace when 4 or 5
fingers are touching.

Signed-off-by: Daniel Kurtz <djkurtz@chromium.org>
Acked-by: Chase Douglas <chase.douglas@canonical.com>

Change-Id: I58f39a14e808ba865e5c39dfdf72e7b6da112395
BUG=chromium-os:19094
TEST=builds clean; evtest to confirm raw events
     confirm motion, 2f-tap, 2f-scroll, using xf86-input-cmt
Reviewed-on: http://gerrit.chromium.org/gerrit/4259
Reviewed-by: Andrew de los Reyes <adlr@chromium.org>
Reviewed-by: Daniel Kurtz <djkurtz@chromium.org>
Tested-by: Daniel Kurtz <djkurtz@chromium.org>