genirq: implement suspend/resume tasks
This patch is essentially a backport of changes that went in to 3.0 linux
kernel. We want, in particular, changes where irq controllers with no
wakeup source configuration are supported in the generic irq code.
This relieves the individual chip drivers from masking the non-wakeup
irqs in suspend and unmasking them in resume.

This will also ease the migration of irq_controllers for 3.0 merge.

Change-Id: I3b55e4c06c298563e3c4cc76330b283f04bdb19e
Signed-off-by: Abhijeet Dharmapurikar <adharmap@codeaurora.org>