CHROMIUM: tegra: usb_phy: Wake up ULPI PHY when device is connected
During run-time, .bus_suspend callback in ehci-tegra.c will be called if
device is in idle state or removed. Gobi-fw test tool will simulate a
device "disconnected", wait for time to be longer than 2 seconds, and then
"connected".
In this 2-second waiting (default autosuspend timer time), .bus_suspend
callback function will be called to put bus into low power state.
When device is simulated as "connected", we expect this detection will wake up
bus. But it's not. Because of the following.

1. Clock to PHY is stopped. So "connected" event is not detected.
2. Controller clock (link layer) is stopped. So "connected" event doesn't
notify system.
3. WKCN (Wake on Connect, bit 20) of PORTSC register is set by upper layer
driver, but it's cleared by current code.

BUG=chrome-os-partner:4050
TEST=Ran "gobi-fw activate D3600-STSUVH-1576" in three cases.
1. Wait 5 seconds before next run.
2. Wait 2 seconds before next run.
3. No waiting before next run.

Change-Id: I918acecaf717fcb0ec683a667ae076070ad39e3f
Signed-off-by: Jim Lin <jilin@nvidia.com>
Reviewed-on: http://gerrit.chromium.org/gerrit/6291
Reviewed-by: Grant Grundler <grundler@chromium.org>