CHROMIUM: seccomp_filter: remove "skip" from copy and add drop helper
seccomp_filters_copy() had a skip argument because it used to use a
preallocated array of filters.  Using skip made it easier to drop entries
without reorganizing them prior to copy.  The new use of btrees gets rid of
this complexity and allows a drop to be done relatively painlessly just after
the copy (when needed).

Signed-off-by: Will Drewry <wad@chromium.org>

BUG=chromium-os:14496
TEST=see dep cl

Change-Id: I0dc19998eb1b0463125e7f53f2f1e441246a0d90
Reviewed-on: http://gerrit.chromium.org/gerrit/6326
Reviewed-by: Sonny Rao <sonnyrao@chromium.org>
Tested-by: Will Drewry <wad@chromium.org>