mmc: msm_sdcc: Fix issue with SDCC runtime suspend/resume framework
Below is the problematic scenario that this patch fixes -

The SDCC host is runtime suspended and the system suspend
has began. Before starting suspend of all devices, the usage
counter is incremented by power management framework. Now, MMC
block device is suspended as part of devices suspend but before
suspend of SDCC host, suspend is aborted because of some other
device. In this case, resume is called for all suspended devices
but the counter is not yet decremented. So when MMC block resume
is called and when it tries to resume the host, it fails. This is
because the current msmsdcc_enable() code resumes only when device
usage counter is 0. This code is now changed to use
pm_runtime_suspended() to determine when to resume the device.

Change-Id: Ib4290c24715c3fdfbfbdb23305c814c28abfaa74
Signed-off-by: Sahitya Tummala <stummala@codeaurora.org>