mmc: msm_sdcc: Fix issue with SDCC runtime suspend/resume framework
Below is the problematic scenario that this patch fixes -

The SDCC host is runtime suspended and the system suspend
has began. Before starting suspend of all devices, the usage
counter is incremented by power management framework. Now, MMC
block device is suspended as part of devices suspend but before
suspend of SDCC host, suspend is aborted because of some other
device. In this case, resume is called for all suspended devices
but the counter is not yet decremented. So when MMC block resume
is called and when it tries to resume the host, it fails. This is
because the current msmsdcc_enable() code resumes only when device
usage counter is 0. This code is now changed to use
pm_runtime_suspended() to determine when to resume the device.

Signed-off-by: Sahitya Tummala <stummala@codeaurora.org>
(cherry picked from commit 68338bf199fb1700e559db7366cffce80df4af5f)

Change-Id: I2b236632bbad2dfd18a5b70dea2c4998f137274e
CRs-fixed: 302776
Signed-off-by: Shruthi Krishna <skrish@codeaurora.org>