regulator: pm8921-regulator: correct 100 mA LVS pull down setting
The pull down enable bit of the 100 mA LVS control register was set
the same way as those in LDO and SMPS regulators.  I.e. set to 1 to
enable the pull down and set to 0 to disable it.  However, the pull
down bit meaning for 100 mA LVS's is inverted.  Correct this in the
driver.

Change-Id: I5a9c0593f2f0c23624b7103321a518b49b9357ea
Signed-off-by: David Collins <collinsd@codeaurora.org>