msm_serial_hs_lite: Enable clocks when adding the port
Enable the GSBI bus clock when calling uart_add_one_port in
the probe function to allow a clean transition from the
early console to the regular console. The TTY layer may try
to configure the port before the early console became
deactivated, which will prematurely disable the clocks and
cause a lockup.

Change-Id: I63ab57e0a134605026d7281afbba8175a355bca9
Signed-off-by: Stepan Moskovchenko <stepanm@codeaurora.org>