msm: charm: Always set GPIO to the documented value
Early versions of charm SURFs required AP2MDM_KPDPWR be
set to 0 to power on the external modem. This was changed with
later hardware versions which require that AP2MDM_KPDPWR be set
to 1 to power on the external modem. The old hardware is no
longer used, so drop support for it and always set AP2MDM_KPDPWR
to 1 for powering on.

Change-Id: I6f4b75891d59a68b42090d22846ae059afd6a2c9
CRs-Fixed: 298499
Signed-off-by: Laura Abbott <lauraa@codeaurora.org>