ASoC: WCD9310: Precharge LDO_H and CFILT before hs polling
Fixes the issue of button press interrupts coming upon
inserting the headset

The problem was that the microphone bias voltage was not
ramping up quickly enough during the start of button
polling. This caused microphone voltage readings to be
inaccurate. By precharging the microphone bias supplies,
the microphone bias ramps up quickly enough to accurately
measure the microphone voltage.

Change-Id: Ib9594e3200355e7823acb5c63b0ea9d5646b9a3c
CRs-fixed: 299410, 302071
Signed-off-by: Brad Rubin <brubin@codeaurora.org>