power: pm8921-charger: remove efficiency improvements
Writing 0x8C to 0x209 makes the pmic charging efficient but causes
a voltage to appear on usb after it is removed. This makes the
phone never detect usb removal.  Remove writing 0x8C to 0x209.

Also remove writing 0xF1 to 0x209. This was thought to be helping
with the issue, but it did not help.

Change-Id: Icddab0aa6c04127bce50e2cc26a7e50da321ebb8
Signed-off-by: Abhijeet Dharmapurikar <adharmap@codeaurora.org>