usb: gsmd: Free usb ep req if req staus is cable disconnection
USB ep request is being added to the list and being freed later
w/o deleting from the list when usb cable is disonnected. This
is causing memory corruption and kernel panic in other function
drivers. Fix the issue by not adding the request to list during
cable disconnection

Change-Id: I527ccfc98a4916a0afff8b1335b5e58038164708
CRs-Fixed: 301905
Signed-off-by: Vamsi Krishna <vskrishn@codeaurora.org>