cfg80211/nl80211: Indicate roaming feature capability to userspace.
When the rssi of the current AP drops, both wpa_supplicant and the
firmware may do a background scan to find a better AP and try to
associate. Since firmware based roaming is faster, inform
wpa_supplicant to avoid roaming and let the firmware decide to
roam if necessary.

For fullmac drivers like ath6kl, it is just enough to provide the
ESSID and the firmware will decide on the BSSID. Since it is not
possible to do pre-auth during roaming for fullmac drivers, the
wpa_supplicant needs to completely disconnect with the old AP and
reconnect with the new AP. This consumes lot of time and it is
better to leave the roaming decision to the firmware.

Signed-off-by: Vivek Natarajan <nataraja@qca.qualcomm.com>

BUG=chrome-os-partner:5421
TEST=matfunc for regression; test w/ modified wpa_supplicant to verify bgscan is now disabled

Change-Id: I6ff27fab12e64cc5f564f753210f4f2a22916748
Reviewed-on: http://gerrit.chromium.org/gerrit/6852
Reviewed-by: Kalle Valo <kvalo@qca.qualcomm.com>
Reviewed-by: Paul Stewart <pstew@chromium.org>
Tested-by: Sam Leffler <sleffler@chromium.org>
(cherry picked from commit 074cb45d8a731821d0eaa46f7885e8edbce9f926)