perf: Skip ctx iteration if task has no events
If a task has been previously profiled, perf-core doesn't
completely destroy the perf_event_context unless the task exits.
As a simpler solution, this patch solves the problem where every time
such a task switches in it causes perf-core to needlessly reprogram
the currently running events from a cpu_ctx.

The problem manifests as an exponential increase in counter output.

Change-Id: Ieec0bbc55a780ba5ac85a19a189ca14696c959a2
Signed-off-by: Ashwin Chaugule <ashwinc@codeaurora.org>