msm: kgsl: Debugfs knob to configure wait timeout of adreno device
There are times when the adreno device is heavily loaded. Here, a
predetermined wait timeout value will not be useful. The configurability
introduced with this change will help in understanding the behaviour and
to tune the value. Will also be helpful in quickly identifying problems
that arise when the irqs don't fire.

Change-Id: I828b596b08a762f093095d35bba0cb645a0f6143
Signed-off-by: Ranjhith Kalisamy <ranjhith@codeaurora.org>