CHROMIUM: destroy_workqueue: lock around checking cwqs flushed
In destroy_workqueue, we need to take cwq->gcwq->lock to ensure we are
seeing fresh values for nr_active and delayed_works.

(A version of this against Linus' HEAD was accepted by the maintainer of
the workqueue subsystem and has been sent to akpm and stable@kernel.org.
I expect this to be similar, if not identical, to the backport to this
kernel version.)

BUG=chrome-os-partner:5866
TEST=Adhoc; a small test-case chain-working module hangs instead of panicing.

Signed-off-by: ttuttle <ttuttle@chromium.org>

Change-Id: I73fe4906524d0fa31cd9dd93e5c0f336508ffff6
Reviewed-on: http://gerrit.chromium.org/gerrit/7350
Reviewed-by: Vincent Palatin <vpalatin@chromium.org>
Tested-by: ttuttle <ttuttle@chromium.org>