CHROMIUM: destroy_workqueue: lock around checking cwqs flushed
In destroy_workqueue, we need to take cwq->gcwq->lock to ensure we are
seeing fresh values for nr_active and delayed_works.

(A version of this against Linus' HEAD was accepted by the maintainer of
the workqueue subsystem and has been sent to akpm and stable@kernel.org.
I expect this to be similar, if not identical, to the backport to this
kernel version.)

BUG=chrome-os-partner:5866
TEST=Adhoc; a small test-case chain-working module hangs instead of panicing.

Signed-off-by: ttuttle <ttuttle@chromium.org>

Reviewed-on: http://gerrit.chromium.org/gerrit/7350
Reviewed-by: Vincent Palatin <vpalatin@chromium.org>
Tested-by: ttuttle <ttuttle@chromium.org>
(cherry picked from commit 9cfa753bec6e7f8dbe4cc031ed28ba6a28380cd5)

Change-Id: Ib09566c6c6577a5a497c828f181816ac2d623bce
Reviewed-on: http://gerrit.chromium.org/gerrit/8035
Reviewed-by: ttuttle <ttuttle@chromium.org>
Reviewed-by: Vincent Palatin <vpalatin@chromium.org>
Tested-by: Jason Glasgow <jglasgow@chromium.org>