msm: calculate where the movable zone should be on 8960
For DMM to work properly, the memory which we intend
to power off must be placed in the "movable zone"
so that all pages in this memory may be migrated out.

Calculate the start and size of the movable zone -- for
the initial 8960 configuration this coincides with
the maximum size of unstable memory, in future
configurations this might be smaller (so that there
is enough memory in the "normal" zone where unmovable
pages will be placed).

Change-Id: I1509ee9db7078aba7e58cbfb9a495146d511721d
Signed-off-by: Larry Bassel <lbassel@codeaurora.org>