usb_storage: realtek_cr: workaround possible freeze when activating auto-delink
When there is no card present, the power management feature called
"auto-delink" cuts the USB connection from the card device to the host
controller.
The "auto-delink" is activated by sending a specially crafted CBW URB,
but once this URB is sent, the device cuts quickly the USB link under
the host controller feet. So we never get the answer to the subsequent
CSW packet, most of the time the controller code detects the
disconnection and completes the URB transaction with an error (EPROTO),
but sometimes we get stuck waiting for the URB completion.
This change adds a 250ms timeout on the USB transaction to ensure a proper
fallback if we are stuck waiting for the completion.

Signed-off-by: Vincent Palatin <vpalatin@chromium.org>

BUG=chromium-os:18236
TEST=suspend/resume an Alex machine (by actioning the lid) multiple
times while playing Netflix as described in the bug.

Change-Id: Ib67e412beac3d377ba5e6afeb22ea60e09976446
Reviewed-on: http://gerrit.chromium.org/gerrit/8438
Reviewed-by: Sameer Nanda <snanda@chromium.org>
Reviewed-by: Puneet Kumar <puneetster@chromium.org>
Tested-by: Vincent Palatin <vpalatin@chromium.org>
Commit-Ready: Vincent Palatin <vpalatin@chromium.org>