CHROMIUM: usb_storage: realtek_cr: workaround possible freeze when activating auto-delink
When there is no card present, the power management feature called
"auto-delink" cuts the USB connection from the card device to the host
controller.
The "auto-delink" is activated by sending a specially crafted CBW URB,
but once this URB is sent, the device cuts quickly the USB link under
the host controller feet. So we never get the answer to the subsequent
CSW packet, most of the time the controller code detects the
disconnection and completes the URB transaction with an error (EPROTO),
but sometimes we get stuck waiting for the URB completion.
This change adds a 250ms timeout on the USB transaction to ensure a proper
fallback if we are stuck waiting for the completion.

Signed-off-by: Vincent Palatin <vpalatin@chromium.org>

BUG=chromium-os:21253
TEST=suspend/resume an Alex machine (by actioning the lid) multiple
times while playing Netflix.

Change-Id: I99dbbd7ae945ddf635a05868501988290a8ccd63
Reviewed-on: http://gerrit.chromium.org/gerrit/8762
Reviewed-by: Olof Johansson <olofj@chromium.org>
Tested-by: Vincent Palatin <vpalatin@chromium.org>
Commit-Ready: Vincent Palatin <vpalatin@chromium.org>