genirq: explicitly mask a freed irq
When an interrupt is freed, the shutdown or the disable callback
is called for that interrupt. These calls might not be implemented
or even if they were, might not mask the interrupt.

Explicitly mask the interrupt when it is freed. If not masked, the
interrupt could trigger, set the pending bit in the irq controller
and cause unnecessary wakeup or exits from idle power collapse.

Change-Id: I5db89b48ac8ed168f3ac0888f9450acf9e27fb17
Signed-off-by: Abhijeet Dharmapurikar <adharmap@codeaurora.org>