CHROMIUM: x86: print out initial usec value calculated from TSC and CPU HZ
This is to give an idea of how much time was spent in firmware and kernel
prior to timekeeping initialization in the kernel.  We print out TSC now,
but printing out usec is more portable across architectures.  We can
remove the raw TSC print once the scripts have been changed over to use
usec.

BUG=chromium-os:4900
TEST=boot on an x86 platform, verify that there is a line in the
kernel log saying "Initial usec timer"

Change-Id: I0a90643f383a937821f8a6439e994ffdaf43d454
Signed-off-by: Sonny Rao <sonnyrao@chromium.org>
Reviewed-on: http://gerrit.chromium.org/gerrit/7584
Reviewed-by: Olof Johansson <olofj@chromium.org>