mmc: core: Increase write data timeout value for bad SDHC cards
SD2.0 specification says the max. timeout for writes is 250msecs
for SDHC cards. But when the card is almost full and there are too few
free sectors, the garbage collectors inside SD card  may take more time
and hence delaying the write more than 250msecs. This typically happens
with cards that don't adhere to specification. Based on this the timeout
is increased to 300msecs but still for some cards this is not sufficient.

Few SanDisk class4 SD cards violate the spec. and may take more time than
expected when the card is full. So increase the data timeout to 800msecs,
otherwise we might see file system corruption due to loss of important FS
metadata in timeouts.

Signed-off-by: Sujith Reddy Thumma <sthumma@codeaurora.org>
(cherry picked from commit 5ab2437f40010e5c717a48ac06f6401311ca6a7c)

Change-Id: I5a9ba9c2f5d6ee0cd4f03cac46df684920661587
Signed-off-by: Oguri, Srivalli <oguri@codeaurora.org>