mm: make physical memory offline work
In the .38 version, the platform specific physical
offline returns the number of bytes offlined, so
a value of 0 indicates an error, not success as in
the previous version. Make sure that the memory
for the original memory resource nodes is not
freed via kfree, as this memory was obtained
from alloc_bootmem very early in the system's life.

Change-Id: I7f404ef64e9aaab67ed40ed96b5b006129933eb0
Signed-off-by: Larry Bassel <lbassel@codeaurora.org>