spi_qsd: Fix for SPI Operational State Invalid error
This error is reproted randomly when the SPI core is put
into RUN state and occurs when the ACPU clock is low.
When the timer expires, we check again to ensure that the
STATE_VALID bit is set before returning.

CRs-fixed: 304672
Signed-off-by: Harini Jayaraman <harinij@codeaurora.org>
(cherry picked from commit 52e065ba3d86977b59937693ac7e85836cf4eca8)

Change-Id: Iabb539ce48776d4c37e9f4e97daac76caa5f9526
Signed-off-by: Raj Kushwaha <rajk@codeaurora.org>