drm/i915: Check eDP power when doing aux channel communications
Verify that the eDP VDD is on, either with the panel being on or with
the VDD force-on bit being set.

This demonstrates that in many instances, VDD is not on when needed,
which leads to failed EDID communications.

Signed-off-by: Keith Packard <keithp@keithp.com>
(cherry picked from commit 9b984daec45632c4c1ef6e628dca4d2bc8f544ed)

Signed-off-by: Stéphane Marchesin <marcheu@chromium.org>
Change-Id: I6baa82ff4b2d0176268c304f1c94f0ee7453a045
Reviewed-on: https://gerrit.chromium.org/gerrit/13285
Reviewed-by: Olof Johansson <olofj@chromium.org>