CHROMIUM: config: enable ip6_tunnel config
Add ip6_tunnel module to kernel config.  This module is needed
for Dualstack-Lite (RFC 6333) support, for clients operating in
an IPv6-only network.

In theory, any platform could operate in an IPv6-only network.
Dualstack-Lite is one of a very small number of transition
technologies that provides IPv4 access in this situation.

Signed-off-by: Midori Kato <aoimidori27@chromium.org>
Signed-off-by: Erik Kline <ek@google.com>

BUG=chromium-os:20219
TEST=modprobe ip6_tunnel

Change-Id: I8a068773d5380da8765f0db8a58970474e463224
Reviewed-on: http://gerrit.chromium.org/gerrit/7973
Tested-by: Midori Kato <aoimidori27@chromium.org>
Reviewed-by: Yoshiki IGUCHI <yoshiki@chromium.org>