UPSTREAM: ath9k: load noise floor from history after the full chip reset
Currently during the full reset, the nf calibration is always
restarted from the defaults. The noise floor history buffers are
never be used again after the scan and ath reset. This patch
ensures that nf histories are always be used that helps to
improve the signal quality on congested environment

Signed-off-by: Rajkumar Manoharan <rmanohar@qca.qualcomm.com>
Signed-off-by: John W. Linville <linville@tuxdriver.com>
Signed-off-by: Paul Stewart <pstew@chromium.org>

BUG=chrome-os-partner:4886
TEST=MatFunc+SecMat (for regression), shuttle run.

Change-Id: I635b942f352db8806d05d83ce5c3dc1c6a81719c
Reviewed-on: http://gerrit.chromium.org/gerrit/8455
Tested-by: Paul Stewart <pstew@chromium.org>
Reviewed-by: Paul Stewart <pstew@chromium.org>