power: pm8921-bms: calculate real_fcc
real_fcc is the remaining capacity in the battery when the charger
driver indicates that the battery is fully charged. This value needs
to be calculated when charging ends with a fully charged battery.

The userspace can read this value via module parameter and preserve it
between restarts.

The real reason for calculating this is to indicate how much the battery
capacity has dropped w.r.t  the profiled battery data use this deviation
in calculations.

Change-Id: I4f6e2fa8b059bf31b17b0f8261f1bdcecc1b6779
Signed-off-by: Abhijeet Dharmapurikar <adharmap@codeaurora.org>