msm: 8960: add voting for SDCC DFAB clock
SDCC hclk is derived from DFAB (dytona fabric) clock which makes
SDCC hclk rate same as the DFAB clock rate.

SDCC driver should be voting for DFAB clock rate when it's active
(not suspended). This voting is already present for 8660 but missing
on 8960. So this change makes the SDCC driver vote for DFAB clock
rate to be 64MHz when active. If there is no voting in place from SDCC
driver and USB is not connected then DFAB clock runs at 27 MHz.

CRs-fixed: 308068
Change-Id: I2eca13c5a7837b9bb14dd2e06407e6231b061448
Signed-off-by: Subhash Jadavani <subhashj@codeaurora.org>