msm: msm_bus: Optimize clock calls and abstract RPM functionality
Clock calls are grouped per request to reduce the number of requests
sent to the RPM for setting clock rates. Clocks are disabled when rate
is set to 0 and enabled when rate goes from 0 to non-zero value.
RPM functionality is abstracted in a different file for structuring the
code in facilitating above optimization.

Change-Id: I737e7095cd61771e8a7331cab08689ef097e6764
Signed-off-by: Gagan Mac <gmac@codeaurora.org>

Signed-off-by: Raj Kushwaha <rajk@codeaurora.org>