gpu: ion: Add O_DSYNC flag to generated file
Ion uses the O_DSYNC flag of a file to indicate an uncached mapping.
When generating a file for mmaping, the new file needs to have the
same O_DSYNC flag as the parent (the file of the ioctl open call)
to indicate if the mapping should be cached or uncached.

Change-Id: If25f45f4e1530aeed481ce76b731a24d3d3a395a
Signed-off-by: Laura Abbott <lauraa@codeaurora.org>