ASoC: msm: 8960: make FLUID ANC mic bias arrangement as default
CDP and FLUID, though both support ANC headset, micbias arrangement
for ANC microphones are done differently. To be able to detect CDP
and FLUID, EEPROM must be programmed and it is difficult to enforce.
Instead, the decision is to only test ANC on FLUID and make FLUID
ANC mic bias arrangement as default

Change-Id: Ifcdb47014c8463f7a2192e0741dc151ee98390b5
Signed-off-by: Patrick Lai <plai@codeaurora.org>